memory to float

+++ main.go
@@ -16,7 +16,7 @@ var (
 	FlagCPU                    = flag.Duration("c", 0, "Interval for CPU waste")
 	FlagCPUduration            = flag.Duration("d", 0, "Min duration for each CPU waste")
 	FlagCPUpercent             = flag.Float64("p", 100.0, "CPU load percentage")
-	FlagMemory                 = flag.Int("m", 0, "GiB of memory waste")
+	FlagMemory                 = flag.Float64("m", 0, "GiB of memory waste")
 	FlagNetwork                = flag.Duration("n", 0, "Interval for network speed test")
 	FlagNetworkConnectionCount = flag.Int("t", 10, "Set concurrent connections for network speed test")
 )
